+++ apps/public-docsite/src/SiteDefinition/SiteDefinition.pages/Controls/Controls.pages.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { INavPage, LoadingComponent } from '@fluentui/react-docsite-components/lib/index2';
 import {
   controlsPagesWeb,
+  controlsPagesWc,
   controlsPagesIos,
   controlsPagesAndroid,
   controlsPagesMac,
@@ -21,6 +22,7 @@ export const ControlsPages: INavPage = {
     require.ensure([], require => cb(require<any>('../../../pages/Overviews/ControlsPage/ControlsPage').ControlsPage)),
   platforms: {
     web: controlsPagesWeb,
+    webcomponents: controlsPagesWc,
     ios: controlsPagesIos,
     android: controlsPagesAndroid,
     mac: controlsPagesMac,